fixed bug related to messages being over 2000 chars
when given a large number of dice to roll caex would display an odd
message this was due to throwing and catching IllegalArgumentException
when given a bad die input. this would be fine except
TextChannel#sendMessage() also throws this when given a message that is
over 2000 characters. 
in order to correct this, i have created an IllegalDiceFormatException
that I throw and catch instead